split query parameter listener into a validator + a listener (need to split tests as well)
